+++ app/src/main/java/com/jmarser/weatherapp_java/main/presenter/WeatherPresenter.java
@@ -2,5 +2,5 @@ package com.jmarser.weatherapp_java.main.presenter;
 
 public interface WeatherPresenter {
 
-    void getWeatherForLocation(long latitude, long longitude);
+    void getWeatherForLocation(double latitude, double longitude);
 }
